feat!: remove msw and read base url from environment variables (#1026)
* feat!: remove msw and read base url from environment variables

* chore: re-generate lockfile

* chore: run prettier

* ci: add sample environment variables copy script to nodejs action

* ci: remove copy script

* ci: add env

* ci: move env to run step

* ci: remove env

* ci: set node version to 18

* ci: add env

* chore: prettier
