+++ .env.sample
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_BASE_URL=https://apigo.afetharita.com
