+++ .github/workflows/nodejs.yml
@@ -1,9 +1,9 @@
 name: Node.JS CI - Check Lint - Code Format - Build AfetHarita
 on:
   push:
-    branches: [main, development]
+    branches: [ main, development ]
   pull_request:
-    branches: [main, development]
+    branches: [ main, development ]
 permissions:
   contents: read
 jobs:
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [16.x]
+        node-version: [ 18.x ]
     steps:
       - uses: actions/checkout@v3
       - name: Setup Node.js ${{ matrix.node-version }}
@@ -31,6 +31,8 @@ jobs:
         run: |
           yarn prettier
       - name: Build AfetHarita
+        env:
+          NEXT_PUBLIC_BASE_URL: https://apigo.afetharita.com
         if: ${{ success() }}
         run: |
           yarn build
