+++ .prettierignore
@@ -5,4 +5,3 @@ node_modules
 public
 package.json
 .github
-mocks
