+++ package.json
@@ -74,7 +74,6 @@
     "eslint-config-next": "13.1.6",
     "husky": "^8.0.3",
     "lint-staged": "13.1.2",
-    "msw": "^1.0.0",
     "prettier": "^2.8.3",
     "typescript": "4.9.5"
   },
