+++ src/components/UI/Drawer/components/channels/babala/FeedChannelBabala.tsx
@@ -1,6 +1,6 @@
 import Typography from "@mui/material/Typography";
 import Chip from "@mui/material/Chip";
-import { FeedChannelBabalaProps } from "../../types";
+import { FeedChannelBabalaProps } from "@/types";
 import { Divider } from "@mui/material";
 import { capitalize, isNaN } from "@/utils/helpers";
