+++ src/components/UI/Map/LeafletMap.tsx
@@ -1,5 +1,5 @@
 import Map from "@/components/UI/Map/Map";
-import { EVENT_TYPES, MarkerData } from "@/mocks/types";
+import { EVENT_TYPES, MarkerData } from "@/types";
 import {
   useDevice,
   useIsDrawerOpen,
