+++ src/mocks/areasResponse.json
@@ -1,109 +1,145 @@
 {
-	"count": 35,
-	"results": [{
-		"id": 13390,
-		"loc": [36.63779, 36.21024]
-	}, {
-		"id": 13279,
-		"loc": [36.63779, 36.21024]
-	}, {
-		"id": 13220,
-		"loc": [36.8866002, 36.2244442]
-	}, {
-		"id": 13169,
-		"loc": [36.8866002, 36.2244442]
-	}, {
-		"id": 13096,
-		"loc": [36.8866002, 36.2244442]
-	}, {
-		"id": 13066,
-		"loc": [36.88634440000001, 36.21425869999999]
-	}, {
-		"id": 13028,
-		"loc": [36.8866002, 36.2244442]
-	}, {
-		"id": 12931,
-		"loc": [36.88634440000001, 36.21425869999999]
-	}, {
-		"id": 12778,
-		"loc": [36.6403472, 36.213862]
-	}, {
-		"id": 12662,
-		"loc": [36.6403472, 36.213862]
-	}, {
-		"id": 12385,
-		"loc": [36.6403472, 36.213862]
-	}, {
-		"id": 12316,
-		"loc": [36.6403472, 36.213862]
-	}, {
-		"id": 12081,
-		"loc": [36.7680772, 36.231215]
-	}, {
-		"id": 12057,
-		"loc": [36.8726431, 36.2131433]
-	}, {
-		"id": 11812,
-		"loc": [36.7709008, 36.2239639]
-	}, {
-		"id": 11803,
-		"loc": [36.954006, 36.2428475]
-	}, {
-		"id": 10518,
-		"loc": [36.840677, 36.197491]
-	}, {
-		"id": 10460,
-		"loc": [36.7680772, 36.231215]
-	}, {
-		"id": 9890,
-		"loc": [36.8726431, 36.2131433]
-	}, {
-		"id": 8900,
-		"loc": [36.7680772, 36.231215]
-	}, {
-		"id": 8240,
-		"loc": [36.8726431, 36.2131433]
-	}, {
-		"id": 8014,
-		"loc": [36.7560915, 36.2192411]
-	}, {
-		"id": 7652,
-		"loc": [36.7560915, 36.2192411]
-	}, {
-		"id": 7637,
-		"loc": [36.7107057, 36.2245253]
-	}, {
-		"id": 6963,
-		"loc": [36.7680772, 36.231215]
-	}, {
-		"id": 6882,
-		"loc": [36.6197988, 36.22656509999999]
-	}, {
-		"id": 5620,
-		"loc": [36.7680772, 36.231215]
-	}, {
-		"id": 5328,
-		"loc": [36.8351618, 36.2264794]
-	}, {
-		"id": 3856,
-		"loc": [36.639914, 36.2152611]
-	}, {
-		"id": 3655,
-		"loc": [36.7629998, 36.2112793]
-	}, {
-		"id": 3462,
-		"loc": [36.6117115, 36.2278245]
-	}, {
-		"id": 3070,
-		"loc": [36.8889076, 36.230703]
-	}, {
-		"id": 1768,
-		"loc": [36.8322922, 36.2050613]
-	}, {
-		"id": 1468,
-		"loc": [36.761078, 36.2200511]
-	}, {
-		"id": 805,
-		"loc": [36.7575155, 36.200711]
-	}]
+  "count": 35,
+  "results": [
+    {
+      "id": 13390,
+      "loc": [36.63779, 36.21024]
+    },
+    {
+      "id": 13279,
+      "loc": [36.63779, 36.21024]
+    },
+    {
+      "id": 13220,
+      "loc": [36.8866002, 36.2244442]
+    },
+    {
+      "id": 13169,
+      "loc": [36.8866002, 36.2244442]
+    },
+    {
+      "id": 13096,
+      "loc": [36.8866002, 36.2244442]
+    },
+    {
+      "id": 13066,
+      "loc": [36.88634440000001, 36.21425869999999]
+    },
+    {
+      "id": 13028,
+      "loc": [36.8866002, 36.2244442]
+    },
+    {
+      "id": 12931,
+      "loc": [36.88634440000001, 36.21425869999999]
+    },
+    {
+      "id": 12778,
+      "loc": [36.6403472, 36.213862]
+    },
+    {
+      "id": 12662,
+      "loc": [36.6403472, 36.213862]
+    },
+    {
+      "id": 12385,
+      "loc": [36.6403472, 36.213862]
+    },
+    {
+      "id": 12316,
+      "loc": [36.6403472, 36.213862]
+    },
+    {
+      "id": 12081,
+      "loc": [36.7680772, 36.231215]
+    },
+    {
+      "id": 12057,
+      "loc": [36.8726431, 36.2131433]
+    },
+    {
+      "id": 11812,
+      "loc": [36.7709008, 36.2239639]
+    },
+    {
+      "id": 11803,
+      "loc": [36.954006, 36.2428475]
+    },
+    {
+      "id": 10518,
+      "loc": [36.840677, 36.197491]
+    },
+    {
+      "id": 10460,
+      "loc": [36.7680772, 36.231215]
+    },
+    {
+      "id": 9890,
+      "loc": [36.8726431, 36.2131433]
+    },
+    {
+      "id": 8900,
+      "loc": [36.7680772, 36.231215]
+    },
+    {
+      "id": 8240,
+      "loc": [36.8726431, 36.2131433]
+    },
+    {
+      "id": 8014,
+      "loc": [36.7560915, 36.2192411]
+    },
+    {
+      "id": 7652,
+      "loc": [36.7560915, 36.2192411]
+    },
+    {
+      "id": 7637,
+      "loc": [36.7107057, 36.2245253]
+    },
+    {
+      "id": 6963,
+      "loc": [36.7680772, 36.231215]
+    },
+    {
+      "id": 6882,
+      "loc": [36.6197988, 36.22656509999999]
+    },
+    {
+      "id": 5620,
+      "loc": [36.7680772, 36.231215]
+    },
+    {
+      "id": 5328,
+      "loc": [36.8351618, 36.2264794]
+    },
+    {
+      "id": 3856,
+      "loc": [36.639914, 36.2152611]
+    },
+    {
+      "id": 3655,
+      "loc": [36.7629998, 36.2112793]
+    },
+    {
+      "id": 3462,
+      "loc": [36.6117115, 36.2278245]
+    },
+    {
+      "id": 3070,
+      "loc": [36.8889076, 36.230703]
+    },
+    {
+      "id": 1768,
+      "loc": [36.8322922, 36.2050613]
+    },
+    {
+      "id": 1468,
+      "loc": [36.761078, 36.2200511]
+    },
+    {
+      "id": 805,
+      "loc": [36.7575155, 36.200711]
+    }
+  ]
 }
